refactor: renaming package step3 to step5
