--- src/main/java/step3/util/CarNamesSeparator.java
@@ -1,16 +1,0 @@
-package step3.util;
-
-import java.util.Arrays;
-import java.util.List;
-
-public class CarNamesSeparator {
-    private static final String SEPARATOR = ",";
-
-    private CarNamesSeparator() {
-        throw new AssertionError();
-    }
-
-    public static List<String> split(final String carNames) {
-        return Arrays.asList(carNames.split(SEPARATOR));
-    }
-}
