--- src/main/java/step3/worker/domain/strategy/MovableStrategy.java
@@ -1,5 +1,0 @@
-package step3.worker.domain.strategy;
-
-public interface MovableStrategy {
-    boolean isMove();
-}
