--- src/main/java/step3/worker/interfaces/render/ViewRender.java
@@ -1,5 +1,0 @@
-package step3.worker.interfaces.render;
-
-public interface ViewRender {
-    String render();
-}
