--- src/main/java/step3/worker/interfaces/render/WinnerResultViewRender.java
@@ -1,18 +1,0 @@
-package step3.worker.interfaces.render;
-
-import java.util.List;
-
-import static step3.worker.interfaces.render.ViewString.WINNER_ANNOUNCEMENT;
-
-public class WinnerResultViewRender implements ViewRender {
-    private final List<String> winnerNames;
-
-    public WinnerResultViewRender(final List<String> winnerNames) {
-        this.winnerNames = winnerNames;
-    }
-
-    @Override
-    public String render() {
-        return String.join(",", winnerNames) + WINNER_ANNOUNCEMENT;
-    }
-}
