+++ src/main/java/step5/client/view/output/Printer.java
@@ -1,0 +1,11 @@
+package step5.client.view.output;
+
+public class Printer {
+    public static void println(final String x) {
+        System.out.println(x);
+    }
+
+    public static void print(final String x) {
+        System.out.print(x);
+    }
+}
