+++ src/main/java/step5/util/CarNamesSeparator.java
@@ -1,0 +1,16 @@
+package step5.util;
+
+import java.util.Arrays;
+import java.util.List;
+
+public class CarNamesSeparator {
+    private static final String SEPARATOR = ",";
+
+    private CarNamesSeparator() {
+        throw new AssertionError();
+    }
+
+    public static List<String> split(final String carNames) {
+        return Arrays.asList(carNames.split(SEPARATOR));
+    }
+}
