+++ src/main/java/step5/worker/domain/Snapshot.java
@@ -1,0 +1,20 @@
+package step5.worker.domain;
+
+
+import java.util.List;
+
+public class Snapshot {
+    private final List<Car> cars;
+
+    private Snapshot(final List<Car> cars) {
+        this.cars = cars;
+    }
+
+    public static Snapshot of(final List<Car> carRacingResults) {
+        return new Snapshot(carRacingResults);
+    }
+
+    public List<Car> getCars() {
+        return cars;
+    }
+}
