+++ src/main/java/step5/worker/domain/strategy/MovableStrategy.java
@@ -1,0 +1,5 @@
+package step5.worker.domain.strategy;
+
+public interface MovableStrategy {
+    boolean isMove();
+}
