+++ src/main/java/step5/worker/interfaces/render/ViewRender.java
@@ -1,0 +1,5 @@
+package step5.worker.interfaces.render;
+
+public interface ViewRender {
+    String render();
+}
